+++ src/types/data.d.ts
@@ -132,6 +132,7 @@ interface DependencyGraph {
         };
         pretty_name: string;
         tags: string[];
+        manual: string;
     };
 }
 
@@ -319,6 +320,10 @@ interface WorkflowState {
             name: string;
             id: string;
         };
+    };
+    informationPanelData: {
+        open: boolean;
+        currentAnalysis: string;
     };
     hoveredAnalysis: string | null;
     executions: DatawolfExecutionFile[];